Move file into lib dir
