+++ pages/_app.tsx
@@ -2,7 +2,7 @@ import { QueryClientProvider, Hydrate } from '@tanstack/react-query';
 import type { AppProps } from 'next/app';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { useState } from 'react';
-import queryClientInstance from './queryClient';
+import queryClientInstance from '../lib/queryClient';
 import globalStyles from '@/styles/globals';
 import typography from '@/styles/typography';
